+++ src/app/pages/demo-components/filter-components/reset-all/reset-all.component.ts
@@ -1,3 +1,4 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { MdsFilterResetComponent, MdsHightlightPrismModule } from 'medes-ui';
 
@@ -6,12 +7,13 @@ import { MdsFilterResetComponent, MdsHightlightPrismModule } from 'medes-ui';
   templateUrl: './reset-all.component.html',
   styleUrls: ['./reset-all.component.scss'],
   standalone: true,
-  imports: [MdsHightlightPrismModule, MdsFilterResetComponent]
+  imports: [CommonModule, MdsHightlightPrismModule, MdsFilterResetComponent]
 })
 export class ResetAllComponent {
   selected = { category: ['shoes'] };
 
-componentdemo = `<!-- Reset Filter Component -->
+componentdemo = `
+<!-- Reset Filter Component -->
 <mds-filter-reset [class]="'btn btn-primary'" [content]="'Reset All'" [(filterSelected)]="selected"></mds-filter-reset>`;
 
 }