+++ src/app/pages/demo-components/prismjs-highlight/prismjs-highlight.component.ts
@@ -10,14 +10,17 @@ import { MdsHightlightPrismModule } from 'medes-ui';
   imports: [CommonModule, MdsHightlightPrismModule]
 })
 export class DemoPrismjsHighlightComponent {
-  importmodule = `import { MdsHightlightPrismModule } from 'medes-ui';
+  importmodule = `
+import { MdsHightlightPrismModule } from 'medes-ui';
+
 @NgModule({
   declarations: [ ... ],
   imports: [
     MdsHightlightPrismModule
   ]
 })`;
-importcompnent = `<mds-hightlight-prism [code]="importmodule" language="typescript"></mds-hightlight-prism>`;
+importcompnent = `
+<mds-hightlight-prism [code]="importmodule" language="typescript" [heightHighlight]="500"></mds-hightlight-prism>`;
 
   // constructor() { }