+++ src/app/pages/demo-utils/aray-utils/aray-utils.component.ts
@@ -10,8 +10,10 @@ import { MdsHightlightPrismModule } from 'medes-ui';
   imports: [CommonModule, MdsHightlightPrismModule]
 })
 export class ArrayUtilsComponent {
-import = `import { MdsArrayUtils } from 'medes-ui';`;
-demo = `MdsArrayUtils.countUniqueValues(["Web","Web","Web","Web","iOS","iOS","Android","Android","Android"])
+import = `
+import { MdsArrayUtils } from 'medes-ui';`;
+demo = `
+MdsArrayUtils.countUniqueValues(["Web","Web","Web","Web","iOS","iOS","Android","Android","Android"])
 // output {Web: 4, iOS: 2, Android: 3}
 
 MdsArrayUtils.sumValues([5,3,4,56,6,7]);