+++ src/app/pages/demo-utils/number-utils/number-utils.component.ts
@@ -10,8 +10,10 @@ import { MdsHightlightPrismModule } from 'medes-ui';
   imports: [CommonModule, MdsHightlightPrismModule]
 })
 export class NumberUtilsComponent {
-import = `import { MdsNumberUtils } from 'medes-ui';`;
-demo = `MdsNumberUtils.shortNumber(15320); // output 15k
+import = `
+import { MdsNumberUtils } from 'medes-ui';`;
+demo = `
+MdsNumberUtils.shortNumber(15320); // output 15k
 MdsNumberUtils.shortNumber(15320, ['rb','jt','mil','tr']); // output 15rb
 
 MdsNumberUtils.toCurrency(15320, 'USD'); // output $15,320