+++ src/app/pages/demo-utils/string-utils/string-utils.component.ts
@@ -10,8 +10,10 @@ import { MdsHightlightPrismModule } from 'medes-ui';
   imports: [CommonModule, MdsHightlightPrismModule]
 })
 export class StringUtilsComponent {
-import = `import { MdsStringUtils } from 'medes-ui';`;
-demo = `MdsStringUtils.lowerCase('This Is String') // output this is string
+import = `
+import { MdsStringUtils } from 'medes-ui';`;
+demo = `
+MdsStringUtils.lowerCase('This Is String') // output this is string
 MdsStringUtils.upperCase('this is string') // output THIS IS STRING
 MdsStringUtils.titleCase('this is string') // output This Is String
 MdsStringUtils.pascalCase('this is string') // output ThisIsString