+++ src/app/shared/components/aside/aside.component.ts
@@ -64,4 +64,11 @@ export class AsideComponent {
   ];
   // constructor() { }
 
+  resetScroll(){
+      document.getElementById('article').scrollTo({
+        top: 0,
+        behavior: 'smooth'
+      });
+  }
+
 }